:recycle: refactor: 페이지에 있어야 할 내용은 page로 옮기기
